improving query text to sql performance

+++ app/api/api-modules/dataSource/dataSource.service.ts
@@ -1,7 +1,9 @@
 import DataSourceModel from './dataSource.model';
 import DataSourceConnectorService from './dataSourceConnector.service';
 import { IDataSourceCreate } from './interfaces/IDataSourceCreate';
-import { createDataSchemasForDataSource } from '../dataSchema/dataSchema.service';
+import { createDataSchemasForDataSource, deleteSchemasForDataSource } from '../dataSchema/dataSchema.service';
+import { createDataSchemaEmbeddings, deleteEmbeddingsForDataSource } from '../chroma/chroma.service';
+import { fulfilledPromiseValues } from '../../utils/promise.server';
 
 export const createDataSource = async (dataSourceData: IDataSourceCreate) => {
   const dataSource = await DataSourceModel.create(dataSourceData);
@@ -10,7 +12,8 @@ export const createDataSource = async (dataSourceData: IDataSourceCreate) => {
   const dataBaseSchema = await connectorService.getSchema();
   const tableSchemas = await getTableSchemas(connectorService, dataBaseSchema);
   await connectorService.disconnect();
-  await createDataSchemasForDataSource(dataSource, dataBaseSchema, tableSchemas);
+  const dataSchemas = await createDataSchemasForDataSource(dataSource, dataBaseSchema, tableSchemas);
+  await createDataSchemaEmbeddings(dataSource, dataSchemas);
   return dataSource;
 }
 
@@ -22,15 +25,25 @@ export const getDataSources = () => {
   return DataSourceModel.findMany();
 }
 
-export const deleteDataSource = (id: number) => {
-  return DataSourceModel.deleteById(id);
+export const deleteDataSource = async (id: number) => {
+  const dataSource = await DataSourceModel.findById(id);
+  if (!dataSource) {
+    throw new Error('Data source not found');
+  };
+  await deleteEmbeddingsForDataSource(dataSource);
+  await deleteSchemasForDataSource(dataSource);
+  await DataSourceModel.deleteById(id);
 }
 
 const getTableSchemas = async (connectorService: DataSourceConnectorService, dataBaseSchema: Array<Record<string, any>>) => {
   const tableSchemas: Array<Record<string, string>> = [];
-  await Promise.all(dataBaseSchema.map(async (table: Record<string, any>) => {
-    const tableSchema = await connectorService.getTableSchema(table.tableName);
-    tableSchemas.push({ name: table.tableName, schema: JSON.stringify(tableSchema) });
+  const results = await Promise.allSettled(dataBaseSchema.map(async (table: Record<string, any>) => {
+    const schema = await connectorService.getTableSchema(table.tableName);
+    const samples = await connectorService.getTableRowSamples(table.tableName);
+    tableSchemas.push({ 
+      name: table.tableName, 
+      schema: JSON.stringify({ schema, samples })
+    })
   }));
-  return tableSchemas;
+  return fulfilledPromiseValues(results);
 }
